requirement.txt

djangorestframework==3.15.2
django-crispy-forms
crispy-tailwind
ChatterBot
gunicorn
